feat(s60-e2): mhc_post_layer_mix bf16 register residency 0.73x -> 0.77x
- keep x/residual loads in bf16 registers, widen to fp32 only at contraction
  (bf16->fp32 lossless, semantics identical); halves register pressure
- +37% over epoch-1 (0.561x -> 0.770x), correctness 3/3 PASS
- full exploration documented: tl.dot catastrophic (0.019x, M=K=4 << tensor-core
  tile), explicit 16-MAC slower, split-h worse; tl.sum + bf16 + BLOCK_H=1024 is
  the hand-written ceiling

+++ kernels/track1-triton/mhc_post_layer_mix/s60/epoch2/triton_mhc_post_layer_mix_e2_001.py
@@ -46,13 +46,16 @@ def _post_layer_mix_kernel(
         h_idx = h0 + hb
         h_mask = h_idx < h
 
-        xv = tl.load(x_ptr + ab * h + h_idx, mask=h_mask, other=0.0).to(tl.float32)
+        # bf16 kept in registers (halved register pressure vs immediate fp32 widen),
+        # widened to fp32 only at the contraction; bf16->fp32 is lossless so
+        # semantics are identical to base's fp32 path within tolerance.
+        xv = tl.load(x_ptr + ab * h + h_idx, mask=h_mask, other=0.0)
         r_off = ab * m * h + k[:, None] * h + h_idx[None, :]
-        r_block = tl.load(residual_ptr + r_off, mask=h_mask[None, :], other=0.0).to(tl.float32)
+        r_block = tl.load(residual_ptr + r_off, mask=h_mask[None, :], other=0.0)
 
-        acc = tl.sum(comb[:, :, None] * r_block[:, None, :], axis=0)  # [m, BLOCK_H]
+        acc = tl.sum(comb[:, :, None] * r_block[:, None, :].to(tl.float32), axis=0)  # [m, BLOCK_H]
 
-        out = acc + post[:, None] * xv[None, :]
+        out = acc + post[:, None] * xv[None, :].to(tl.float32)
         tl.store(
             out_ptr + ab * m * h + k[:, None] * h + h_idx[None, :],
             out.to(tl.bfloat16),
